backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Tatiana Belova <tanja5676@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
python-dotenv = "^1.0.1"
django-phonenumbers = "^1.0.1"
django-phonenumber-field = "^8.0.0"
django-colorfield = "^0.11.0"
djoser = "^2.2.3"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
pytils = "^0.4.1"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
